Clean up cron delete returns + replace audit logger calls

+++ app/application/use_cases/cron/delete_cron_job.py
@@ -5,10 +5,14 @@ class DeleteCronJob:
         self.cron_scheduler = cron_scheduler
 
     def execute(self, job_id):
+        """
+        Returns:
+            Bool: True if successfully deleted, false otherwise.
+        """
         job = self.cron_repository.get(job_id)
-        try:
-            self.cron_scheduler.delete(job)
-            self.cron_repository.delete(job_id)
-        except Exception as e:
-            print(e)
 
+        # If we can't delete from the system, don't delete from DB.
+        if self.cron_scheduler.delete(job):
+            return self.cron_repository.delete(job_id)
+
+        return False
